// lib/shopify/types.ts
export type Money = {
  amount: string;
  currencyCode: string;
};
// ...
export type Cost = {
  subtotalAmount: Money;
  totalAmount: Money;
  totalTaxAmount: Money;
}